Finalized Bytecode & Syntax
With the exception of objects (classes) and arrays (vectors) bytecode
generation and syntax evaluation is complete. The generators must be
finished in order to actually compile code. Generation is in the early
iteration stages and needs to be optimized. Code can almost be run but
there are some integration issues with the compiler.