Equivalent Operator (==)
Added the Equivalent Operator along with all the fixes and support
needed for it. While it still needs more support for doubles and strings
at this point in time. Also added a few new helper functions for it
along with some new machine codes such as SETE support for
optimizations.