Improved Extensions and Jumps
A massive reworking of Jumps to allow for larger then 126 bytes in jump
length up to 2.1 billion bytes. Extensions also improved to have better
debugging and features. Also added support for Debug and Release
versions of extensions along with the ability to extend Crystal Clear
with non Clear Export .dlls.