Array Constructor
You can now construct arrays using the array brackets. It's not yet
possible to access arrays yet using the bracket operator. There is
currently a heap problem due to a double free. Strings are conflicting
with pointers for now and will most likely have to be refactored,
speeding up and simplifying garbage collection.