Removed Reference Counting
Reference counting removed since it was simply too slow, had to much
overhead, and leaked due to the famous circular reference problem. The
code is stable (and faster for all the wrong reasons) but it leaks
almost all of the dynamic memory it uses. I need to replace all the
malloc, and calloc calls with a proper garbage collector which I've
already started working on.